+++ src/component/redux/action/types.js
@@ -4,7 +4,6 @@ export const updateInput ="updateInput";
 export const arrayForD3="arrayForD3";
 export const askData ="askData";
 export const chooseCountry ="chooseCountry";
-
 /*
 export const setUoIndex
 export const*/
